Mock Refactoring 1. Platform and Storage mocks are kept in separte files. 2. mocks are moved to test-config folder. 3. Storage mock is refined.
