+++ test-config/mocks/mocks.ts
@@ -1,0 +1,43 @@
+import { StatusBar } from '@ionic-native/status-bar';
+import { SplashScreen } from '@ionic-native/splash-screen';
+
+export class NavMock {
+ 
+  public pop(): any {
+    return getStubPromise();
+  }
+ 
+  public push(): any {
+    return getStubPromise();
+  }
+ 
+  public getActive(): any {
+    return {
+      'instance': {
+        'model': 'something',
+      },
+    };
+  }
+ 
+  public setRoot(): any {
+    return true;
+  }
+}
+
+export class StatusBarMock extends StatusBar {
+  styleDefault() {
+    return;
+  }
+}
+
+export class SplashScreenMock extends SplashScreen {
+  hide() {
+    return;
+  }
+}
+
+function getStubPromise() {
+    return new Promise((resolve: Function) => {
+      resolve();
+    });
+ }
