+++ test-config/mocks/platform.mock.ts
@@ -1,34 +1,3 @@
-import { StatusBar } from '@ionic-native/status-bar';
-import { SplashScreen } from '@ionic-native/splash-screen';
-
-export class NavMock {
- 
-  public pop(): any {
-    return getStubPromise();
-  }
- 
-  public push(): any {
-    return getStubPromise();
-  }
- 
-  public getActive(): any {
-    return {
-      'instance': {
-        'model': 'something',
-      },
-    };
-  }
- 
-  public setRoot(): any {
-    return true;
-  }
-}
-
-export class StorageMock {
-  public ready(): any {
-    return getStubPromise();
-  }
-}
 export class PlatformMock {
   public ready(): Promise<string> {
     return new Promise((resolve) => {
@@ -93,20 +62,3 @@ export class PlatformMock {
     return document['activeElement'];
   }
 }
-export class StatusBarMock extends StatusBar {
-  styleDefault() {
-    return;
-  }
-}
-
-export class SplashScreenMock extends SplashScreen {
-  hide() {
-    return;
-  }
-}
-
-function getStubPromise() {
-    return new Promise((resolve: Function) => {
-      resolve();
-    });
- }
